patch-0.1.25
RD-13 - The Discussions main page is not showing the authors of the threads, Although author information is being returned by the getThreadList API call.
RD-14 - On the 'topics' page, clicking on a forum under 'forums index' does not update the list of topics displayed (although the title at the top of the page is updated, and it looks like the correct getTopicList API call is being made)
RD-15 - On the 'Topics for Forum' page, in addition to the topics, a list of forums is also displayed in the main list. The correct getTopicList API call is being made, and the correct list is coming back. Not sure why the forum list is being appended to the topic list.
RD-16 - For the 'Forums Index' at the right-hand side of the page, the column labels should be 'Forum' and 'Topics' (not 'topics' and 'threads')
RD-17 - On the 'Create Thread' page, once an image has been added with the 'upload image' option, clicking on the 'x' does not delete it. I don't see a call being made to the 'deletePostImage ' API call.
RD-18 - On Mars (which is using the live discussions data server), the white background runs out on the 'Topics in forum: Deep Sky', when all topics (21) are displayed.
RD-20 - In discussion boards: if an image is submitted as part of a reply, it does not show up in the display of a thread.
RD-23 - On the discussion boards, when creating a new thread, the lists of forums and topics should be left-aligned (the same way lists are left-aligned when creating new Community Content).