patch-1.0.0
- PRC2 Discussion Board New Thread:
When adding a moderate amount of content to the New Thread content panel, the scroll bar won't scroll far enough to reveal the "Submit New Thread" button. User had to use keyboard to scroll down the page. Likewise in the upward direction....and Discussion Board Thread Size:
If the content of a thread goes beyond a certain (moderate) level, it results in two scroll bars - one on the thread panel and the browser window.
- Adding additional language support for compatibility with internet explorer